requirements.txt
astroid
beautifulsoup4
certifi
chardet
click
colorama
Flask
Flask-Cors
gunicorn
idna
isort
itsdangerous
Jinja2
lazy-object-proxy
lxml
MarkupSafe
mccabe
numpy
Pillow
pymongo
requests
six
urllib3
Werkzeug
wrapt
